Add per-channel Atom feed with 10 most recent messages
The feed contains the web version of the messages, so including base64-encoded
images and without the subscribe/unsubscribe verbiage.

This is related to #14 except that it uses Atom instead of RSS; Atom should be
supported by all clients as well, and is slightly more thoroughly standardized
in RFCs and such. Feed readers can be pointed either at the channel page or at
the feed URL directly.

Tried with Liferea and Thunderbird.